Move URLRequest setting functions into a separate file
